+++ examples/postgres-soft-deletion/main.rs
@@ -108,7 +108,7 @@ async fn main() {
         .expect("Todo item should be created");
     println!("My first todo created {:?}", todo);
 
-    let first_todo = Todo::get_by_id(&pool, todo.id).await.unwrap();
+    let first_todo = Todo::get_by_id(&pool, &todo.id).await.unwrap();
     match first_todo {
         Some(ref item) => println!("First todo item is {:?}", item),
         None => println!("Todo item does not exist for the id {0}", todo.id),
@@ -121,14 +121,14 @@ async fn main() {
         .await
         .expect("Item should be updated");
 
-    let check_updated_item = Todo::get_by_id(&pool, todo.id).await.unwrap();
+    let check_updated_item = Todo::get_by_id(&pool, &todo.id).await.unwrap();
     match check_updated_item {
         Some(ref item) => println!("Updated item is {:?}", item),
         None => println!("Todo item does not exist for the id {0}", todo.id),
     }
 
     check_updated_item.unwrap().delete(&pool).await.unwrap();
-    let deleted_todo = Todo::get_by_id(&pool, todo.id).await.unwrap();
+    let deleted_todo = Todo::get_by_id(&pool, &todo.id).await.unwrap();
     match deleted_todo {
         Some(ref item) => println!("Todo item still exists What??? / {:?}", item),
         None => println!(